+++ docs/jazzy.yml
@@ -141,6 +141,7 @@ custom_categories:
   - name: User Feedback
     children:
       - FeedbackViewController
+      - FeedbackSubtypeViewController
       - FeedbackViewControllerDelegate
       - FeedbackItem
       - FeedbackType
